Replace generic feature step with two specific steps
When the application only sorted one way, we only needed a step
called "My records are sorted".  In the upcoming release, the
app will sort in more than one way.  Hence, replacing this step
with two that specify sorting in different ways.

Also, it is expected that the record manager will choose the
sorting method, hence the "I choose..." form.

* This commit breaks the "record manager views records sorted
  by last name descending" feature of the previous release, since
  new code (Session#show_by_last_name_descending) is referenced but
  not implemented.