Reorganize features
Instead of maintaining a feature called "Record manager starts sorter"
and another that combines different kinds of sorting in "Record manager
views sorted records", it makes sense to drop the first and split up
the second since the application is not meant to be interactive.

This also allows us to think of our sortings differently so that
we can imagine them as instantiated according to the chosen sort
method and polymorphically treated from there.