Collapse Given steps for each feature
It is not easy or desirable to maintain state between these
two steps, so it is better to consider them as a single step.

This makes more sense considering that the application is run
as a command line utility with both the sorting method and the
records given as arguments at initial execution. Both are chosen
at the same time.