Sort using Enumerable#sort_by
...to improve readability.  #sort_by is more succinct than #sort
when the sorting is explicit.  However, #sort is more efficient,
so we may return to #sort if we expect to handle many records at once.

* Side-effect: Using #sort_by exposed a flaw in the Sorter spec.
Enumerable#sort doesn't evaluate its block when it is called on
a collection with only one element, while #sort_by does. In other
words, these throw errors:

[nil].sort_by { |a| a + 1 }
[nil, nil].sort { |a, b| a + 1 <=> b + 1 }

but this does NOT throw an error:

[nil].sort { |a, b| a + 1 <=> b + 1 }