Add Record class
Sorter is currently tasked with the responsibilities of both parsing/
interpreting records and sorting them in different orders. It would
be better if each record knew about its own internal organization so
that Sorter could focus on sorting, hence the need for a Record class.

This commit adds the Record class and specs for its existing behavior,
but it does not integrate Record into the application at all.  The
next step will be to rewrite the specs for classes that will collaborate
with record, and use these specs to drive the rewrites of those classes.