Add feature: Handle unrecognized instruction
Currently the record-parser executable throws an error when
an unrecognized sorting instruction is given. This feature describes
how unrecognized instruction is handled.

A new class, RecordParser::Application, is referenced. This class
will be the new home of the code that is currently kept in the
executable, and it will have unit specs.