Add RecordParser::Application class and spec
RecordParser::Application::run is the single point of entry into
the application. Therefore, the executable bin/record-parser needs
only to call that method with the inputs from the user.

This commit includes specs for RecordParser::Application. The Cucumber
collaborator method FakeOutput#puts had to be edited so that it returns
nil, as does #puts normally.