Handle bad filename and notify user
Added a scenario to the "Enters bad input" feature to cover the
case when the file cannot be read.

Because checking whether the file can be read and acting accordingly
required two calls to Application's private method ::unsorted_records,
the value of the first call was memoized to prevent two calls of ARGF#read
(since the second ARGF#read would be nil). However, the memoizing
broke the unit and feature tests, since the Application class now
carried state from test to test. It was decided that it was better
to instantiate Application rather than work around this strange
behavior.

Ruby: "NO SINGLETON FOR YOU!"
Me: "Le Wistful Sigh."