Correct filename given when file unreadable
Previously, when multiple files were given as input and one of them
was unreadable, the failure message did not correctly capture the
bad filename. Instead, the filename given in the failure message
was th next file in the sequence, which may not have been unreadable
at all.

This was happening because the call to ARGF#filename occurred after
the failure to open, and when there were multiple files, #filename's
value had already changed to the next filename.

The bug is fixed in this commit by capturing the bad filename directly
from the message of the error being rescued. Specs are provided, as
is a patch to the features' FakeFile collaborator.