Validate sorting instruction before processing file
Fixes issue #1.

Previously, the CommandLineInterface read the file before considering
the instruction, which allows the situation described in the issue;
essentially, the CommandLineInterface interpreted the filename as
an instruction and lack of any other filename as standard input, and
since the instruction had not been found invalid, the interface
continued to expect standard input.

As of this commit, the CommandLineInterface checks the instruction first,
so that this confusion doesn't arise. A spec example is added for this
behavior, and the feature specs are altered to explicitly pass in each
of the good instructions when testing for a bad file, so that the bad
file scenario isn't implicitly letting the instruction be empty.

It is still possible to use standard input, if one enters a recognized
sorting instruction. In this case, the application works as expected,
sorting the input according to the method chosen.